face-attendance-system/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
customtkinter==5.2.2
Pillow==10.1.0
numpy==1.24.3
